requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.0
django-appconf==1.0.4
django-base64field==1.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-email-verification==0.0.7
django-getenv==1.3.2
django-phonenumber-field==5.0.0
django-user-accounts==3.0.2
django-uuid-upload-path==1.0.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.10
Pillow==7.2.0
progress==1.5
psycopg2==2.8.6
pyparsing==2.4.6
python-dotenv==0.14.0
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.0
toml==0.10.1
urllib3==1.25.8
webencodings==0.5.1
